+++ hrp/cmd/convert.go
@@ -7,7 +7,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/httprunner/httprunner/hrp/internal/convert"
+	"github.com/httprunner/httprunner/v4/hrp/internal/convert"
 )
 
 var convertCmd = &cobra.Command{