+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 **go version**
 
 - fix: step request elapsed timing should contain ContentTransfer part
+- fix #1288: unable to go get httprunner v4
 
 **python version**